### 2019-02-28 #1288 @whedon
[PRE REVIEW]: MicrobiomeR:  An R Package for Simplified and Standardized Microbiome Analysis Workflows
**Submitting author:** @grabear (<a href="http://orcid.org/0000-0001-8096-0180">Robert Gilmore</a>)
**Repository:** <a href="https://github.com/vallenderlab/MicrobiomeR" target ="_blank">https://github.com/vallenderlab/MicrobiomeR</a>
**Version:** 0.5.1
