### 2019-02-28 #1288 @danielskatz
👋 @grabear - I don't see the paper part of your submission.  Take a look at the [author submission guidelines](https://joss.readthedocs.io/en/latest/submitting.html), specifically [What should my paper contain?](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) and [Example paper and bibliography](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography)
