### 2019-02-28 #1288 @grabear
Hi, I put the paper in a separate folder under inst/paper/ @danielskatz
