### 2019-02-28 #1288 @danielskatz
thanks - I see some issues with the figure, the table, and the references.  You can make changes to the paper.md and paper.bib, then use the same command I used (as a new comment here) to regenerate until you are satisfied with the generated proof.
