### 2019-02-28 7845d7c
Attempt image fix.
