+++ inst/paper/paper.md
@@ -34,7 +34,10 @@ Previous open source microbiome packages available for R [@r-core] (e.g. **vegan
 # Workflow
 
 The paradigm for studying the microbiome follows the steps shown in the figure below.  MicrobiomeR covers the last four steps of this workflow.
+
+<center>
 ![Microbiome Workflow](Microbiome-Workflow.png)
+</center>
 
 MicrobiomeR provides users with raw output files from the NIH’s Nephele pipeline.  This data was generated by the Illumina MiSeq workflow followed by the **Qiime** [@qiime] 16S pipeline for pair ended FASTQ files.  To begin a MicrobiomeR workflow, raw data files are imported into R as a phyloseq object, using the *create_phyloseq()* function.  While discouraged, the data can be manipulated directly using any of phyloseq’s methods.  Next, the data is converted to a taxmap object using the *create_taxmap()* function.
 
