### 2019-02-28 #1288 @grabear
@sdhutchins working on tweaking the paper!
