### 2019-03-01 #1288 @grabear
Part of the problem is that I always protect my master and dev-master branches, so it makes it a bit more complicated for me to edit.  I can always change the permissions temporarily, but that's not ideal.
