### 2019-03-01 7518b30
Updated table 1 in the paper.
