### 2019-03-01 #1288 @lpantano
sorry @grabear that is not working with two of the conditions (other branch and other location), hopefully, somebody who knows the backend infrastructure can help. I will reach somebody if @arfon doesn't chime in.

At least the figure now. seems right. :)

Thanks!
