### 2019-03-01 #1288 @grabear
@lpantano It seems to be working now!  It's finding the paper, so no complaints.  :)
