### 2019-03-01 fcc6e3f
Updated DOIs and references in .bib file.
