+++ inst/paper/paper.bib
@@ -15,6 +15,7 @@
   volume = {8},
   number = {4},
   year = {2013},
+  doi = {https://doi.org/10.1371/journal.pone.0061217},
   url = {http://dx.plos.org/10.1371/journal.pone.0061217},
 }
 
@@ -54,6 +55,7 @@
   year = {2010},
   volume = {7},
   pages = {335-336},
+  doi = {https://doi.org/10.1038/nmeth.f.303},
   url = {qiime.org},
 }
 
@@ -64,10 +66,17 @@
   year = {2013},
   issue = {41},
   pages = {D590-D596},
+  doi = {https://doi.org/10.1093/nar/gks1219},
 }
 
 @Article{silva_2,
   title = {The SILVA and “All-species Living Tree Project (LTP)” taxonomic frameworks.},
   author = {Yilmaz P and Parfrey LW and Yarza P and Gerken J and Pruesse E and Quast C
-  and Schweer T and Peplies J and Ludwig W and Glöckner FO}
+  and Schweer T and Peplies J and Ludwig W and Glöckner FO},
+  journal = {Nucleic Acids Research},
+  year = {2014},
+  volume = {42},
+  issue = {D1},
+  pages = {D643-D648},
+  doi = {https://doi.org/10.1093/nar/gkt1209},
 }
