### 2019-03-01 50c0e01
Converted table to Latex.
