### 2019-03-01 7de6f12
Reworked markdown table with html page breaks.
