+++ inst/paper/paper.md
@@ -50,16 +50,12 @@ After preprocessing the data, statistical analysis can be conducted using Microb
 One of the key advantages in using MicrobiomeR is its “phyloseq-style” filtering functions.  These functions are heavily dependent upon formatting and validation checkpoints to ensure that the proper data is being manipulated.  In Table 1, we describe the “MicrobiomeR formats” used throughout the package and how they are related to phyloseq, metacoder, and each other.
 
 ### Table 1:
-\begin{table}[]
-\begin{tabular}{|c|c|c|}
-\hline
-Name & Description & Data Tables \\ \hline
-Phyloseq Format & \begin{tabular}[c]{@{}c@{}}A taxmap object that has just been \\ converted from a phyloseq object with the\\ create\_phyloseq() function. The \\ observation tables represent the \\ phyloseq::otu\_table(), phyloseq::tax\_table(),\\  phyloseq::sample\_data(), and \\ phyloseq::phy\_tree().\end{tabular} & \begin{tabular}[c]{@{}c@{}}otu\_table \\ tax\_data \\ sample\_data \\ phy\_tree\end{tabular} \\ \hline
-Raw Format & \begin{tabular}[c]{@{}c@{}}A taxmap object that has been processed \\ with the as\_raw\_format() function.  The \\ otu\_table and tax\_data observation table \\ from the "phyloseq\_format” are renamed.\end{tabular} & \begin{tabular}[c]{@{}c@{}}phyloseq tables \\ otu\_abundance \\ otu\_annotations\end{tabular} \\ \hline
-Basic Format & \begin{tabular}[c]{@{}c@{}}A taxmap object that has been processed \\ with the as\_basic\_format() function.  This \\ format is defined by observation data that has\\  been processed with the metacoder::calc\_*\_()\\  functions.\end{tabular} & \begin{tabular}[c]{@{}c@{}}raw tables \\ taxa\_abundance \\ otu\_proportions \\ taxa\_proportions\end{tabular} \\ \hline
-Analyzed Format & \begin{tabular}[c]{@{}c@{}}A taxmap object that has been processed \\ with the as\_analyzed\_format() function.  \\ This format is defined by observation data \\ that has been processed with the \\ metacoder::compare\_groups() function.\end{tabular} & \begin{tabular}[c]{@{}c@{}}basic tables \\ statistical\_data \\ stats\_tax\_data\end{tabular} \\ \hline
-\end{tabular}
-\end{table}
+| Name | Description | Data Tables |
+|:------------------------------------------------------:|:-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------:|:-----------------------------------------------------------------------------:|
+| Phyloseq Format | A taxmap object that has just been <br>converted from a phyloseq object with the<br> create_phyloseq() function.  The observation<br> tables represent the phyloseq::otu_table(), <br>phyloseq::tax_table(),<br> phyloseq::sample_data(), and<br> phyloseq::phy_tree(). | otu_table<br>  tax_data<br>  sample_data<br>  phy_tree |
+| Raw Format | A taxmap object that has been processed<br> with the as_raw_format() function.  The<br> otu_table and tax_data observation table<br> from the "phyloseq_format” are renamed. | phyloseq tables<br>  otu_abundance<br>  otu_annotations |
+| Basic Format | A taxmap object that has been processed<br> with the as_basic_format() function.  This<br> format is defined by observation data that<br> has been processed with the<br> metacoder::calc_*_() functions. | raw tables<br>  taxa_abundance<br>  otu_proportions<br>  taxa_proportions<br> |
[…]
