+++ inst/paper/paper.md
@@ -52,10 +52,10 @@ One of the key advantages in using MicrobiomeR is its “phyloseq-style” filte
 ### Table 1:
 | Name | Description | Data Tables |
 |:------------------------------------------------------:|:-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------:|:-----------------------------------------------------------------------------:|
-| Phyloseq Format | A taxmap object that has just been <br>converted from a phyloseq object with the<br> create_phyloseq() function.  The observation<br> tables represent the phyloseq::otu_table(), <br>phyloseq::tax_table(),<br> phyloseq::sample_data(), and<br> phyloseq::phy_tree().<br> | otu_table<br>  tax_data<br>  sample_data<br>  phy_tree |
-| Raw Format | A taxmap object that has been processed<br> with the as_raw_format() function.  The<br> otu_table and tax_data observation table<br> from the "phyloseq_format” are renamed.<br> | phyloseq tables<br>  otu_abundance<br>  otu_annotations |
-| Basic Format | A taxmap object that has been processed<br> with the as_basic_format() function.  This<br> format is defined by observation data that<br> has been processed with the<br> metacoder::calc_*_() functions.<br> | raw tables<br>  taxa_abundance<br>  otu_proportions<br>  taxa_proportions<br> |
-| Analyzed Format | A taxmap object that has been processed<br> with the as_analyzed_format()<br> function.  This format is defined by<br> observation data that has been processed<br> with the metacoder::compare_groups()<br> function.<br> | basic tables<br>  statistical_data<br>  stats_tax_data<br> |
+| Phyloseq Format | A taxmap object that has just been <br>converted from a phyloseq object with the<br> create_phyloseq() function.  The observation<br> tables represent the phyloseq::otu_table(), <br>phyloseq::tax_table(),<br> phyloseq::sample_data(), and<br> phyloseq::phy_tree().<br><br> | otu_table<br>  tax_data<br>  sample_data<br>  phy_tree |
+| Raw Format | A taxmap object that has been processed<br> with the as_raw_format() function.  The<br> otu_table and tax_data observation table<br> from the "phyloseq_format” are renamed.<br><br> | phyloseq tables<br>  otu_abundance<br>  otu_annotations |
+| Basic Format | A taxmap object that has been processed<br> with the as_basic_format() function.  This<br> format is defined by observation data that<br> has been processed with the<br> metacoder::calc_*_() functions.<br><br> | raw tables<br>  taxa_abundance<br>  otu_proportions<br>  taxa_proportions<br> |
+| Analyzed Format | A taxmap object that has been processed<br> with the as_analyzed_format()<br> function.  This format is defined by<br> observation data that has been processed<br> with the metacoder::compare_groups()<br> function.<br><br> | basic tables<br>  statistical_data<br>  stats_tax_data<br> |
 
 Validation is performed internally by most of the preprocessing, analysis, and formatting functions, but it can also be done directly by using the *is_\*_format()* functions.  Furthermore, it is encouraged to explore the taxmap objects in order to view observation and taxonomy data.  Other MicrobiomeR utilities include project management tools for creating and organizing output directories for plots, a set of color palette functions based on *grDevices::colorRampPalette()*, and a taxonomic data parsing function called *parse_taxonomy_silva_128()* used to import data annotated by the SILVA [@silva_1] [@silva_2] database as a phyloseq object.
 
