### 2019-03-01 #1288 @grabear
@lpantano @sdhutchins I think I have the table, diagram, and DOIs in a good place.

Thoughts?
