### 2019-03-01 #1288 @lpantano
Thanks!

I think looks good, the only think is that some DOIs have the `https` and other not. Can you  normalize  that?
