### 2019-03-02 #1288 @sdhutchins
I think it looks good, @grabear! I would only say perhaps the diagram's height could be decreased while keeping the same scale. It looks good as it is though.
