### 2019-03-04 21a4341
UPdated DOIs with missing https.
