+++ inst/paper/paper.bib
@@ -29,7 +29,7 @@
   volume = {13},
   pages = {1-15},
   number = {2},
-  doi = {10.1371/journal.pcbi.1005404},
+  doi = {https://doi.org/10.1371/journal.pcbi.1005404},
   url = {http://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1005404},
 }
 
@@ -40,7 +40,7 @@
   year = {2018},
   volume = {7},
   number = {272},
-  doi = {10.12688/f1000research.14013.1},
+  doi = {https://doi.org/10.12688/f1000research.14013.1},
   url = {https://f1000research.com/articles/7-272/v2},
 }
 
