### 2019-03-04 fa67dd2
Updated diagram size.
