### 2019-03-04 #1288 @CosteaPaul
@lpantano this looks like fun :).
Would be happy to help with reviewing, though I have no idea how this is supposed to work.
