### 2019-03-04 #1288 @lpantano
Thank you,  @CosteaPaul !

We have some guidelines here:

https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

For now, I will move this to review, that will close this issue and open a new one where the review process will happen. Our friend whedon, will help you with the requirements to review the paper. When the new issue is open, there will be some guidelines. I will follow up there.

Thanks so much!
