### 2019-03-04 #1288 @lpantano
Hi @grabear,

I see there are some issues with the DOI links , let me know when you  solve this and we can move this to under review.

Thanks!
