+++ inst/paper/paper.bib
@@ -40,7 +40,7 @@
   year = {2018},
   volume = {7},
   number = {272},
-  doi = {https://doi.org/10.12688/f1000research.14013.1},
+  doi = {https://doi.org/10.12688/f1000research.14013.2},
   url = {https://f1000research.com/articles/7-272/v2},
 }
 
