### 2019-03-04 #1288 @grabear
@lpantano All of those DOI's are not missing.  I copied an pasted those during my first check, so Idk why they are considered "Missing" or "may be missing".

I just updated an F1000 DOI, but only because they release a second version of that paper.  I also checked all of the other ones and they are working links..
