### 2019-03-04 #1288 @lpantano
Hi,
Weird, When I click on the PDF link it lands on a broken web page.  Did you try to use the link of the PDF, maybe there is something wrong on the latex code that breaks the link?

`https://doi.org/https://doi.org/10.1038/nmeth.f.303` This url is malformed.

Cheers
