+++ inst/paper/paper.bib
@@ -15,7 +15,7 @@
   volume = {8},
   number = {4},
   year = {2013},
-  doi = {https://doi.org/10.1371/journal.pone.0061217},
+  doi = {10.1371/journal.pone.0061217},
   url = {http://dx.plos.org/10.1371/journal.pone.0061217},
 }
 
@@ -29,7 +29,7 @@
   volume = {13},
   pages = {1-15},
   number = {2},
-  doi = {https://doi.org/10.1371/journal.pcbi.1005404},
+  doi = {10.1371/journal.pcbi.1005404},
   url = {http://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1005404},
 }
 
@@ -40,7 +40,7 @@
   year = {2018},
   volume = {7},
   number = {272},
-  doi = {https://doi.org/10.12688/f1000research.14013.2},
+  doi = {10.12688/f1000research.14013.2},
   url = {https://f1000research.com/articles/7-272/v2},
 }
 
@@ -55,7 +55,7 @@
   year = {2010},
   volume = {7},
   pages = {335-336},
-  doi = {https://doi.org/10.1038/nmeth.f.303},
+  doi = {10.1038/nmeth.f.303},
   url = {qiime.org},
 }
 
@@ -66,7 +66,7 @@
   year = {2013},
   issue = {41},
   pages = {D590-D596},
-  doi = {https://doi.org/10.1093/nar/gks1219},
+  doi = {10.1093/nar/gks1219},
 }
 
 @Article{silva_2,
@@ -78,5 +78,5 @@
   volume = {42},
   issue = {D1},
   pages = {D643-D648},
-  doi = {https://doi.org/10.1093/nar/gkt1209},
+  doi = {10.1093/nar/gkt1209},
 }
