### 2019-03-04 #1288 @grabear
> [ point_right Check article proof page_facing_up point_left ](https://github.com/openjournals/joss-papers/blob/joss.01288/joss.01288/10.21105.joss.01288.pdf)

Not really sure what's going on with whedon's reference check, but the DOIs should be correct now.  Sorry about that!
