### 2019-03-05 #1288 @arfon
> @whedon check references from branch joss-review

I'm afraid @whedon only knows how to check references for the `master` or default branch of your repository.
