### 2019-03-05 #1299 @whedon
[REVIEW]: MicrobiomeR:  An R Package for Simplified and Standardized Microbiome Analysis Workflows
**Submitting author:** @grabear (<a href="http://orcid.org/0000-0001-8096-0180">Robert Gilmore</a>)
**Repository:** <a href="https://github.com/vallenderlab/MicrobiomeR" target ="_blank">https://github.com/vallenderlab/MicrobiomeR</a>
**Version:** 0.6.0

@CosteaPaul , please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
