### 2019-03-05 065b06f
Added contributing.md to README.
