+++ README.md
@@ -39,8 +39,9 @@ Please visit https://microbiomer.vallenderlab.science/ for extensive documentati
     - Alpha Diversity Plot
     - Ordination Plot
 
-## Notes
+## Contributing to `MicrobiomeR`
 
-* Code Style based on Google's R style developed by Hadley Wickham.
+* Our code style based on Google's R style developed by Hadley Wickham.
 * Please note that the `MicrobiomeR` project is released with a [Contributor Code of Conduct](.github/CODE_OF_CONDUCT.md). By contributing to this project, you agree to abide by its terms.
+* View our [contributing](.github/CONTRIBUTING.md) documentation for more details.
 
