### 2019-03-05 bbaa1be
Added deployment of pkgdown site to travis.yml.
