+++ .travis.yml
@@ -28,3 +28,11 @@ r_github_packages:
 after_success:
   - R CMD INSTALL $PKG_TARBALL
   - Rscript -e 'covr::codecov()'
+
+before_deploy:
+  - Rscript -e 'remotes::install_cran("pkgdown")'
+
+deploy:
+  provider: script
+  script: Rscript -e 'pkgdown::deploy_site_github(verbose = TRUE)'
+  skip_cleanup: true
