### 2019-03-05 7f61633
Bumped version to 0.5.2.
