### 2019-03-05 27e87da
Fixed spelling errors.
