+++ man/output_dir.Rd
@@ -35,7 +35,7 @@ A function for generating a consistent file system for data files and visualizat
 of the parameters are set the full path generated by this function will be \emph{root_path}/output
 }
 \details{
-This function is increadibly useful on it's own but also for various other plotting/saving functions within the package.
+This function is incredibly useful on it's own but also for various other plotting/saving functions within the package.
 It helps keep data organized using a standard workflow.
 }
 \examples{
