### 2019-03-05 1a22804
Allow travis to be run for all branches.
