### 2019-03-05 478fa86
Ensure travis does not allow a staged install
