### 2019-03-05 ba2970b
Updated .travis.yml
