+++ .travis.yml
@@ -20,14 +20,14 @@ r:
 r_github_packages:
   - r-lib/covr
 
-after_success:
-  - R CMD INSTALL $PKG_TARBALL --no-staged-install
-  - Rscript -e 'covr::codecov()'
-
 before_deploy:
+  - R CMD INSTALL .
   - Rscript -e 'remotes::install_cran("pkgdown")'
 
 deploy:
   provider: script
   script: Rscript -e 'pkgdown::deploy_site_github(verbose = TRUE)'
   skip_cleanup: true
+
+after_success:
+  - Rscript -e 'covr::codecov()'
