### 2019-03-05 48cc172
Added matrix to .travis.yml
