+++ .travis.yml
@@ -8,26 +8,21 @@ cran: http://cran.rstudio.com
 
 env:
   global:
-    - _R_CHECK_FORCE_SUGGESTS_=false
-    - R_REMOTES_NO_ERRORS_FROM_WARNINGS=true
+    - _R_CHECK_FORCE_SUGGESTS_=FALSE
+    - R_REMOTES_NO_ERRORS_FROM_WARNINGS=TRUE
     - CODECOV_TOKEN=2ff74b01-00e8-4e2f-9cc7-1f90ca17e480
 
-r:
-- oldrel
-- release
-- devel
-
-r_github_packages:
-  - r-lib/covr
-
-before_deploy:
-  - R CMD INSTALL .
-  - Rscript -e 'remotes::install_cran("pkgdown")'
-
-deploy:
-  provider: script
-  script: Rscript -e 'pkgdown::deploy_site_github(verbose = TRUE)'
-  skip_cleanup: true
-
-after_success:
-  - Rscript -e 'covr::codecov()'
+matrix:
+  include:
+  - r: devel
+  - r: release
+    after_success:
+    - Rscript -e 'covr::codecov()'
+    before_deploy:
+    - R CMD INSTALL .
+    - Rscript -e 'remotes::install_cran("pkgdown")'
+    deploy:
+      provider: script
+      script: Rscript -e 'pkgdown::deploy_site_github(verbose = TRUE)'
+      skip_cleanup: true
+  - r: oldrel
