### 2019-03-05 45e7d5d
Add environment variable for staged install.
