+++ .travis.yml
@@ -11,6 +11,7 @@ env:
     - _R_CHECK_FORCE_SUGGESTS_=FALSE
     - R_REMOTES_NO_ERRORS_FROM_WARNINGS=TRUE
     - CODECOV_TOKEN=2ff74b01-00e8-4e2f-9cc7-1f90ca17e480
+    - R_INSTALL_STAGED=FALSE
 
 matrix:
   include:
