### 2019-03-14 #1299 @CosteaPaul
Slightly confused about where exactly i'm supposed to comment on this, so i'll just do it here and i'm sure i'll be told off if i did it wrong.

Running into trouble installing this package:
`install_github("vallenderlab/MicrobiomeR") `
returns:

Error: (converted from warning) unable to access index for repository /src/contrib:
  cannot open URL '/src/contrib/PACKAGES'

devtools 2.0.1
Issue in R 3.5.1, both on Windows and Unbuntu.
