### 2019-03-14 #1299 @grabear
@CosteaPaul This isn't necessarily an issue with our package.

Try using:
`install_github("ropensci/taxa")`
and see if it gives you the same error.

What version of the `remotes` package are you using?  Try upgrading to the latest version if you haven't already.
