### 2019-03-14 #1299 @CosteaPaul
@grabear , ropensci/taxa installs just fine.
Using remotes 2.0.1
