### 2019-03-14 #1299 @grabear
@CosteaPaul Please see the issue that I've created with the remotes repository.  It's linked above. :point_up:

The "solution" was found here https://github.com/r-lib/remotes/issues/251, which links to this:
https://community.rstudio.com/t/travis-build-error-cannot-open-url-src-contrib-packages/18088

If you just do the following:
```r
Sys.setenv(R_REMOTES_NO_ERRORS_FROM_WARNINGS = "true")
remotes::install_github("vallenderlab/MicrobiomeR")
```
