### 2019-03-14 #1299 @grabear
Sorry for the issues, we can update our README to reflect this weird workaround.  Once we put our package on CRAN this issue should not be a problem.
