### 2019-03-17 #1299 @CosteaPaul
@grabear, the little hack seems to solve the install problem.

Ran all from scratch on a docker starting with xubu 18 and the latest R (3.5.3). Installing this takes a while, but runs though. Hefty list of dependencies. Would say overall i spent >2 hours getting this running. (Packages like curl and openssl fail because of lib dependencies that then need to be installed separately).

But now i can start playing with it.

Would actually really appreciate a docker repo that i could just pull and get started directly. Think this is something that would make sense having?
