### 2019-03-18 #1299 @grabear
> Would actually really appreciate a docker repo that I could just pull and get started directly. Think this is something that would make sense having?

I think this would be really nifty!  @CosteaPaul

Unfortunately, our institution has a policy that doesn't permit us to use docker containers on their servers.  I can definitely try to look into it on my personal machine if you think it would be helpful.
