### 2019-03-18 #1299 @CosteaPaul
@grabear it's not something i think is needed, but would be nice to have.
Maybe a singularity container then? (that way you don't have to upset your security ppl).

Anyway, i've ticked all the boxes so as far as i'm concerned this is good to go. Am i supposed to do something else now?
