+++ DESCRIPTION
@@ -13,12 +13,11 @@ Roxygen: list(markdown = TRUE)
 Depends:
     R (>= 3.3.0)
 Remotes: 
-    github::ropensci/taxa,
-    github::grunwaldlab/metacoder,
-    github::joey711/phyloseq,
-    github::microbiome/microbiome,
-    github::r-lib/covr,
-    github::jonclayden/shades
+    ropensci/taxa,
+    grunwaldlab/metacoder,
+    joey711/phyloseq,
+    microbiome/microbiome,
+    jonclayden/shades
 Imports:
     ape,
     biomformat,
