### 2019-03-18 86c025e
Updated ordination_plot documentation.
