### 2019-03-19 c180dfe
Removed CODECOV_TOKEN from .travis.yml.
