### 2019-03-19 038133d
Improved internals of alpha diversity functions.
