### 2019-03-19 ea700df
Update roxygen docs for alpha diversity functions.
