### 2019-03-19 #1299 @lpantano
Hi,

I think all is good. One minor detail in the references is that some of them have initials for the authors and others F. Lastname. Can you normalize the references to have the latter(F Lastname)?

After that, generate another proof, check references, and create a zenodo archive DOI. The metadata needs to be correct, like the title of the version in zenodo should be the title of the paper and author list.

After that, I can start the acceptance process and ping the joss-eics team to have a final look and finalize the acceptance.

Thanks!
