### 2019-03-19 #1299 @lpantano
Thank you, @grabear.

I think there is some misunderstanding, sorry about that.

The Zenodo has to be from your repository, here you have an example of an accepted paper: https://zenodo.org/record/2574400#.XJE3_akpBp8

Cheers
