+++ inst/paper/paper.bib
@@ -80,3 +80,11 @@
   pages = {D643-D648},
   doi = {10.1093/nar/gkt1209},
 }
+
+@Manual{microbiome,
+  title = {microbiome: Tools for microbiome analysis in R},
+  author = {Leo Lahti and Sudarshan Shetty and Tineka Blake and Jarkko Salojarvi},
+  year = {2017},
+  note = {R package version 1.5.28},
+  url = {http://microbiome.github.com/microbiome},
+}
