+++ inst/paper/paper.bib
@@ -83,7 +83,7 @@
 
 @Manual{microbiome,
   title = {microbiome: Tools for microbiome analysis in R},
-  author = {Leo Lahti and Sudarshan Shetty and Tineka Blake and Jarkko Salojarvi},
+  author = {L Lahti and S Shetty and T Blake and J Salojarvi},
   year = {2017},
   note = {R package version 1.5.28},
   url = {http://microbiome.github.com/microbiome},
