### 2019-03-19 #1299 @lpantano
sorry about the issue. Normally I sync Zenodo with my repos and then everytime I make a release a new version in Zenodo is created. If that is easy, go ahead, we can add whatever new Zenodo link to the final reviewed version.
