### 2019-03-19 98ce0f5
Ensure the stat comparisons are in scale with the violin plot height.
