### 2019-03-19 8a34206
Bumped version to 0.5.3
