+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: MicrobiomeR
 Title: Analyze Microbiome Data
-Version: 0.5.2
+Version: 0.6.0
 Authors@R: c(
     person("Robert", "Gilmore", email = "rgilmore@umc.edu", role = c("aut", "cre")),
     person("Shaurita", "Hutchins", email = "shutchins2@umc.edu", role = "aut"))
