### 2019-03-19 a11082b
Updated internals for group.
