+++ NAMESPACE
@@ -1,5 +1,64 @@
 # Generated by roxygen2: do not edit by hand
 
+export("%>%")
+export(agglomerate_taxmap)
+export(alpha_diversity_measures)
+export(alpha_diversity_plot)
+export(as_MicrobiomeR_format)
+export(as_analyzed_format)
+export(as_basic_format)
+export(as_custom_format)
+export(as_phyloseq_format)
+export(as_raw_format)
+export(combination_palette)
+export(correlation_data)
+export(correlation_plot)
+export(correlation_plots)
+export(cov_filter)
+export(create_phyloseq)
+export(create_pub_table)
+export(create_taxmap)
+export(get_color_palette)
+export(heat_tree_parameters)
+export(heat_tree_plots)
+export(is_analyzed_format)
+export(is_basic_format)
+export(is_phyloseq_format)
+export(is_raw_format)
+export(metacoder_comp_func_1)
+export(order_metacoder_data)
+export(ordination_plot)
+export(ordination_plots)
+export(otu_id_filter)
+export(otu_prevalence_filter)
+export(otu_proportion_filter)
+export(output_dir)
+export(parse_taxonomy_silva_128)
+export(permanova)
+export(pick_new_outgroup)
+export(root_by_longest_edge)
+export(root_phyloseq_tree)
+export(sample_id_filter)
+export(save_alpha_diversity_plots)
+export(save_correlation_plots)
+export(save_heat_tree_plots)
+export(save_ordination_plots)
+export(save_stacked_barplots)
+export(save_top_coefficients_barplot)
+export(scico_palette)
+export(stacked_barplot)
+export(stacked_barplots)
+export(taxa_prevalence_filter)
+export(taxon_id_filter)
+export(top_coefficients_barplot)
+export(transformer)
+export(transposer)
+export(treatment_matrix)
+export(validate_MicrobiomeR_format)
+export(viridis_magma_palette)
+export(viridis_palette)
+export(vlookup)
+export(which_format)
 import(vegan)
 importFrom(ape,Ntip)
 importFrom(ape,is.rooted)
