### 2019-03-19 5d2385a
Added travis-ci.com badge.
