### 2019-03-20 9813d3c
Comment out problematic (for travis-ci) tests.
