+++ pkgdown.yml
@@ -1,0 +1,15 @@
+pandoc: '2.2'
+pkgdown: 1.3.0.9000
+pkgdown_sha: 0b2c27d41b869ac31b6a46815940f55b12f6171f
+articles:
+  about: about.html
+  analysis: analysis.html
+  color-palettes: color-palettes.html
+  data-wrangling: data-wrangling.html
+  filtering: filtering.html
+  introduction: introduction.html
+  paper: paper.html
+urls:
+  reference: https://microbiomer.vallenderlab.science//reference
+  article: https://microbiomer.vallenderlab.science//articles
+
