### 2019-03-20 #1299 @grabear
@sdhutchins made a release or two today.  Here is the link to the zenodo for that:
https://zenodo.org/account/settings/github/repository/vallenderlab/MicrobiomeR#

I also used this DOI in the "zenodo mistake" I made earlier.  So we made need to be issued another DOI.
@lpantano
