### 2019-03-20 #1299 @lpantano
Hi,

Thanks for working on this. I found this one: https://zenodo.org/record/2600432#.XJLLXaQpCEc  what seems correct.

But we need to match the title of the zenodo release with the paper (as you did for the mistake). Can you remove the mistake zenodo or change the title and update this one with the paper title.

Probably we need to add more documentation on this matter.

Let me know if you need more help trying to match the title to the paper.

Thanks! almost there.

> On Mar 20, 2019, at 18:08, Rob Gilmore <notifications@github.com> wrote:
>
> @sdhutchins made a release or two today. Here is the link to the zenodo for that:
> https://zenodo.org/account/settings/github/repository/vallenderlab/MicrobiomeR#
>
> I also used this DOI in the "zenodo mistake" I made earlier. So we made need to be issued another DOI.
> @lpantano
>
