### 2019-03-21 #1299 @grabear
Ok so here is the Zenodo link for the paper.  The title should be correct.  I published it as well, so hopefully that solved some other issues.  https://zenodo.org/record/2600432
