### 2019-03-21 #1299 @lpantano
@weldom set 0.6.0 as version
