### 2019-03-21 #1299 @lpantano
Hi @openjournals/joss-eics,

This paper has been accepted. Let us know if we missed something.

Thanks!
