### 2019-03-21 #1299 @sdhutchins
@lpantano just an fyi, we added an additional reference to the paper in the `.bib` file in `v0.60`
