### 2019-03-28 #1299 @lpantano
Just pinging again @openjournals/jose-editors and @arfon in case this was missed. The paper is accepted.
